[Monthly Push] Update skills, work experience, and footer date
Added MySQL, C#, MySQL Workbench, and Visual Studio to the skills list in about.tsx.
Updated project and experience descriptions in work.tsx, including details for Presoft internship and replaced the placeholder image with Presoft.png.
Changed the last updated date in the footer to 3rd September, 2025.

[Copilot]
I forgot to do Food Site, so I might as well get this out of the way for now.